task: move ecommerce events to bundle
