+++ bundles/EcommerceFrameworkBundle/src/Event/Model/CommitOrderProcessorEvent.php
@@ -14,37 +14,33 @@ declare(strict_types=1);
  *  @license    http://www.pimcore.org/license     GPLv3 and PCL
  */
 
-namespace Pimcore\Event\Model\Ecommerce;
+namespace Pimcore\Bundle\EcommerceFrameworkBundle\Event\Model;
 
 use Pimcore\Bundle\EcommerceFrameworkBundle\CheckoutManager\CommitOrderProcessorInterface;
 use Pimcore\Bundle\EcommerceFrameworkBundle\Model\AbstractOrder;
 use Pimcore\Event\Traits\ArgumentsAwareTrait;
 use Symfony\Contracts\EventDispatcher\Event;
 
-class SendConfirmationMailEvent extends Event
+class CommitOrderProcessorEvent extends Event
 {
     use ArgumentsAwareTrait;
 
     protected CommitOrderProcessorInterface $commitOrderProcessor;
 
-    protected AbstractOrder $order;
-
-    protected string $confirmationMailConfig;
-
-    protected bool $skipDefaultBehaviour = false;
+    protected ?AbstractOrder $order = null;
 
     /**
-     * SendConfirmationMailEvent constructor.
+     * CommitOrderProcessorEvent constructor.
      *
      * @param CommitOrderProcessorInterface $commitOrderProcessor
-     * @param AbstractOrder $order
-     * @param string $confirmationMailConfig
+     * @param AbstractOrder|null $order
+     * @param array $arguments
      */
-    public function __construct(CommitOrderProcessorInterface $commitOrderProcessor, AbstractOrder $order, string $confirmationMailConfig)
+    public function __construct(CommitOrderProcessorInterface $commitOrderProcessor, ?AbstractOrder $order, array $arguments = [])
     {
         $this->commitOrderProcessor = $commitOrderProcessor;
         $this->order = $order;
-        $this->confirmationMailConfig = $confirmationMailConfig;
+        $this->arguments = $arguments;
     }
 
     public function getCommitOrderProcessor(): CommitOrderProcessorInterface
@@ -57,7 +53,7 @@ class SendConfirmationMailEvent extends Event
         $this->commitOrderProcessor = $commitOrderProcessor;
     }
 
-    public function getOrder(): AbstractOrder
+    public function getOrder(): ?AbstractOrder
     {
         return $this->order;
     }
@@ -65,25 +61,5 @@ class SendConfirmationMailEvent extends Event
     public function setOrder(AbstractOrder $order): void
     {
         $this->order = $order;
-    }
-
-    public function getConfirmationMailConfig(): string
-    {
-        return $this->confirmationMailConfig;
-    }
-
-    public function setConfirmationMailConfig(string $confirmationMailConfig): void
-    {
-        $this->confirmationMailConfig = $confirmationMailConfig;
-    }
-
-    public function doSkipDefaultBehaviour(): bool
-    {
-        return $this->skipDefaultBehaviour;
-    }
-
-    public function setSkipDefaultBehaviour(bool $skipDefaultBehaviour): void
-    {
-        $this->skipDefaultBehaviour = $skipDefaultBehaviour;
     }
 }
