+++ bundles/EcommerceFrameworkBundle/src/Event/Model/OrderManagerItemEvent.php
@@ -14,7 +14,7 @@ declare(strict_types=1);
  *  @license    http://www.pimcore.org/license     GPLv3 and PCL
  */
 
-namespace Pimcore\Event\Model\Ecommerce;
+namespace Pimcore\Bundle\EcommerceFrameworkBundle\Event\Model;
 
 use Pimcore\Bundle\EcommerceFrameworkBundle\CartManager\CartItemInterface;
 use Pimcore\Bundle\EcommerceFrameworkBundle\Model\AbstractOrderItem;
