+++ bundles/EcommerceFrameworkBundle/src/OrderManager/V7/OrderAgent.php
@@ -32,7 +32,7 @@ use Pimcore\Bundle\EcommerceFrameworkBundle\PaymentManager\StatusInterface;
 use Pimcore\Bundle\EcommerceFrameworkBundle\PaymentManager\V7\Payment\PaymentInterface;
 use Pimcore\Bundle\EcommerceFrameworkBundle\PaymentManager\V7\Payment\RecurringPaymentInterface;
 use Pimcore\Bundle\EcommerceFrameworkBundle\Type\Decimal;
-use Pimcore\Event\Model\Ecommerce\OrderAgentEvent;
+use Pimcore\Bundle\EcommerceFrameworkBundle\Event\Model\OrderAgentEvent;
 use Pimcore\Log\Simple;
 use Pimcore\Logger;
 use Pimcore\Model\DataObject\Concrete;
