+++ bundles/EcommerceFrameworkBundle/src/OrderManager/V7/OrderManager.php
@@ -37,8 +37,8 @@ use Pimcore\Bundle\EcommerceFrameworkBundle\PriceSystem\TaxManagement\TaxEntry;
 use Pimcore\Bundle\EcommerceFrameworkBundle\PricingManager\PriceInfoInterface;
 use Pimcore\Bundle\EcommerceFrameworkBundle\Type\Decimal;
 use Pimcore\Bundle\EcommerceFrameworkBundle\VoucherService\VoucherServiceInterface;
-use Pimcore\Event\Model\Ecommerce\OrderManagerEvent;
-use Pimcore\Event\Model\Ecommerce\OrderManagerItemEvent;
+use Pimcore\Bundle\EcommerceFrameworkBundle\Event\Model\OrderManagerEvent;
+use Pimcore\Bundle\EcommerceFrameworkBundle\Event\Model\OrderManagerItemEvent;
 use Pimcore\File;
 use Pimcore\Logger;
 use Pimcore\Model\DataObject\AbstractObject;
